+++ palette.py
@@ -55,16 +55,19 @@ class Palette:
 
 palette = Palette()
 def test():
+	column = 0
 	for k, v in vars(palette).items():
-		if k=='colors':
-			# Skip the colors dictionary
+		if column >= 3:
+			column = 0
+		if k == 'colors':
+			# Skip self.colors since its just the vars of Palette's attributes
 			continue
-		elif type(v)==dict:
+		elif type(v) == dict:
 			for k2, v2 in v.items():
-				print(f'{v2}{k}[\'{k2}\']')
+				print(f'{v2}{k}[\'{k2}\']', end=' ')
 		else:
-			print(v+k)
-	input('Press ENTER to continue.')
+			print(v+k, end=' ')
+	input('\nPress ENTER to continue.')
 
 if __name__ == '__main__':
 	test()